+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Added `balm.u.Timeline#add_callback/2` add a callback option for Timeline items.
 * Added `balm.u.Tweener#calc_remaining_time/0` calculates the remaining time for the tweener to complete.
+* Added `balm.m.table.split/2`
+* Added `balm.m.table.pop/2`
 
 # 2026.6.4